requirements.txt
aws-cdk-lib>=2.100.0
constructs>=10.3.0
boto3>=1.34.0
pandas>=2.1.0
pyarrow>=14.0.0
requests>=2.31.0

